+++ R/get_tests_group.R
@@ -261,12 +261,10 @@ get_tests_group <-
                                          tzone = base::Sys.timezone())
 
         # Build Output Test Data frame
-        outputDF <-
-          base::cbind(trialInfo, TestTypeData, AthleteData, trialMetrics)
+        outputDF <- base::cbind(trialInfo, TestTypeData, AthleteData, trialMetrics)
 
         # Add Test Meta Data to output data frame
-        outputDF <-
-          dplyr::mutate(outputDF, last_test_time = lastTest, last_sync_time = lastSync)
+        outputDF <- dplyr::mutate(outputDF, last_test_time = lastTest, last_sync_time = lastSync)
 
         # Validate Active Tests
         if (base::isFALSE(includeInactive)) {